Player can move

+++ AdventureGameEngine/interpolation.h
@@ -14,7 +14,7 @@ using namespace std;
 
 double interpolate(vector<double> &xData, vector<double> &yData, double x, bool extrapolate)
 {
-	int size = xData.size();
+	int size = (int)xData.size();
 
 	int i = 0;	// find left end of interval for interpolation
 	if (x >= xData[size - 2])	// special case: beyond right end
